\name{cubic}
\alias{cubic}
\docType{data}
\title{A Data Set for Cgam}
\description{This data set is used for several examples in the cgam package.}
\usage{data(cubic)}
\format{A data frame with 50 observations on the following 2 variables.
\describe{
\item{\code{x}}{The predictor vector.}
\item{\code{y}}{The response vector.}  }  }
\source{STAT640 HW 14 given by Dr. Meyer.  }

\keyword{data set}
